Generate bindings when compiling aws-lc
