+++ .gitignore
@@ -9,7 +9,6 @@ doc/*.html
 doc/doc.css
 bindings/rust/tmp/*
 bindings/rust/generate/target
-bindings/rust/generate/Cargo.lock
 
 util/bot/android_ndk
 util/bot/android_sdk/public
