--- bindings/rust/generate/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "generate"
-version = "0.1.0"
-authors = ["AWS-LC"]
-edition = "2021"
-license = "Apache-2.0"
-# this is an internal tool for generating bindings
-publish = false
-
-[dependencies]
-bindgen = "0.60.1"
-regex = "1"
